+++ wp-content/plugins/core/src/Blocks/Types/Lead_Form/Lead_Form.php
@@ -1,5 +1,4 @@
-<?php
-declare( strict_types=1 );
+<?php declare( strict_types=1 );
 
 namespace Tribe\Project\Blocks\Types\Lead_Form;
 
@@ -17,16 +16,23 @@ class Lead_Form extends Block_Config {
 	public const DESCRIPTION     = 'description';
 	public const CTA             = 'cta';
 
-	public const FORM = 'form';
-
 	public const SECTION_SETTINGS = 's-settings';
 	public const LAYOUT           = 'layout';
-	public const LAYOUT_CENTER    = 'center';
 	public const LAYOUT_LEFT      = 'left';
+	public const LAYOUT_RIGHT     = 'right';
+	public const LAYOUT_BOTTOM    = 'bottom';
 
 	public const WIDTH      = 'width';
 	public const WIDTH_GRID = 'grid';
 	public const WIDTH_FULL = 'full';
+
+	public const BACKGROUND       = 'background';
+	public const BACKGROUND_LIGHT = 'background_light';
+	public const BACKGROUND_DARK  = 'background_dark';
+
+	public const FORM_FIELDS  = 'form_fields';
+	public const FORM_STACKED = 'form_stacked';
+	public const FORM_INLINE  = 'form_inline';
 
 	/**
 	 * Register the block
@@ -41,6 +47,7 @@ class Lead_Form extends Block_Config {
 			'supports'    => [
 				'align'  => false,
 				'anchor' => true,
+				'jsx'    => true,
 			],
 			'example'     => [
 				'attributes' => [
@@ -57,7 +64,6 @@ class Lead_Form extends Block_Config {
 							'url'    => '#',
 							'target' => '',
 						],
-						self::FORM        => 1 //Will work if a site has a form added. Be blank if not.
 					],
 				],
 			],
@@ -72,12 +78,12 @@ class Lead_Form extends Block_Config {
 		// Content Fields
 		//==========================================
 		$this->add_section( new Field_Section( self::SECTION_CONTENT, __( 'Content', 'tribe' ), 'accordion' ) )
-			 ->add_field( new Field( self::NAME . '_' . self::LEAD_IN, [
-					 'label' => __( 'Lead in', 'tribe' ),
-					 'name'  => self::LEAD_IN,
-					 'type'  => 'text',
-				 ] )
-			 )->add_field( new Field( self::NAME . '_' . self::TITLE, [
+			->add_field( new Field( self::NAME . '_' . self::LEAD_IN, [
+					'label' => __( 'Lead in', 'tribe' ),
+					'name'  => self::LEAD_IN,
+					'type'  => 'text',
+				] )
+			)->add_field( new Field( self::NAME . '_' . self::TITLE, [
 					'label' => __( 'Title', 'tribe' ),
 					'name'  => self::TITLE,
 					'type'  => 'text',
@@ -94,44 +100,45 @@ class Lead_Form extends Block_Config {
 					'name'  => self::CTA,
 					'type'  => 'link',
 				] )
-			)->add_field( new Field( self::NAME . '_' . self::FORM, [
-					'label'   => __( 'Form', 'tribe' ),
-					'name'    => self::FORM,
-					'type'    => 'select',
-					'choices' => $this->get_form_options(),
-				] )
 			);
 
 		//==========================================
 		// Setting Fields
 		//==========================================
 		$this->add_section( new Field_Section( self::SECTION_SETTINGS, __( 'Settings', 'tribe' ), 'accordion' ) )
-			 ->add_field(
-				 new Field( self::NAME . '_' . self::LAYOUT, [
-					 'type'            => 'image_select',
-					 'name'            => self::LAYOUT,
-					 'choices'         => [
-						 self::LAYOUT_LEFT   => __( 'Content Left', 'tribe' ),
-						 self::LAYOUT_CENTER => __( 'Content Center', 'tribe' ),
-					 ],
-					 'default_value'   => self::LAYOUT_CENTER,
-					 'multiple'        => 0,
-					 'image_path'      => sprintf(
-						 '%sassets/img/admin/blocks/%s/',
-						 trailingslashit( get_template_directory_uri() ),
-						 self::NAME
-					 ),
-					 'image_extension' => 'svg',
-				 ] )
-			 )->add_field(
-				 new Field( self::NAME . '_' . self::WIDTH, [
-					'type'            => 'image_select',
-					'name'            => self::WIDTH,
+			->add_field(
+				new Field( self::NAME . '_' . self::WIDTH, [
+				'type'            => 'radio',
+				'name'            => self::WIDTH,
+				'label'           => __( 'Container Width', 'tribe' ),
+				'choices'         => [
+					self::WIDTH_GRID => __( 'Grid', 'tribe' ),
+					self::WIDTH_FULL => __( 'Full', 'tribe' ),
+				],
+				'default_value'   => self::WIDTH_GRID,
+				] )
+			)->add_field(
+				new Field( self::NAME . '_' . self::BACKGROUND, [
+					'type'            => 'radio',
+					'name'            => self::BACKGROUND,
+					'label'           => __( 'Background Color', 'tribe' ),
 					'choices'         => [
-						self::WIDTH_GRID => __( 'Grid', 'tribe' ),
-						self::WIDTH_FULL => __( 'Full', 'tribe' ),
+						self::BACKGROUND_LIGHT   => __( 'Light', 'tribe' ),
+						self::BACKGROUND_DARK    => __( 'Dark', 'tribe' ),
 					],
-					'default_value'   => self::WIDTH_GRID,
+					'default_value'   => self::BACKGROUND_LIGHT,
+				] )
+			)->add_field(
+				new Field( self::NAME . '_' . self::LAYOUT, [
+					'type'            => 'image_select',
+					'name'            => self::LAYOUT,
+					'label'           => __( 'Form Layout', 'tribe' ),
+					'choices'         => [
+						self::LAYOUT_LEFT   => __( 'Form Left', 'tribe' ),
+						self::LAYOUT_RIGHT  => __( 'Form Right', 'tribe' ),
+						self::LAYOUT_BOTTOM => __( 'Form Bottom', 'tribe' ),
+					],
+					'default_value'   => self::LAYOUT_BOTTOM,
 					'multiple'        => 0,
 					'image_path'      => sprintf(
 						'%sassets/img/admin/blocks/%s/',
@@ -139,25 +146,26 @@ class Lead_Form extends Block_Config {
 						self::NAME
 					),
 					'image_extension' => 'svg',
-				 ] )
-			 );
+				] )
+			)->add_field(
+				new Field( self::NAME . '_' . self::FORM_FIELDS, [
+					'type'            => 'image_select',
+					'name'            => self::FORM_FIELDS,
+					'label'           => __( 'Form Field Position', 'tribe' ),
+					'choices'         => [
+						self::FORM_STACKED   => __( 'Stacked', 'tribe' ),
+						self::FORM_INLINE    => __( 'Inline', 'tribe' ),
+					],
+					'default_value'   => self::FORM_STACKED,
+					'multiple'        => 0,
+					'image_path'      => sprintf(
+						'%sassets/img/admin/blocks/%s/',
+						trailingslashit( get_template_directory_uri() ),
+						self::NAME
+					),
+					'image_extension' => 'svg',
+				] )
+			);
 	}
-
-	/**
-	 * @return array
-	 */
-	protected function get_form_options() {
-		if ( ! class_exists( 'GFFormsModel' ) ) {
-			return [];
-		}
-		$choices   = [];
-		$choices[] = __( 'Select One', 'tribe' );
-		foreach ( \GFFormsModel::get_forms() as $form ) {
-			$choices[ $form->id ] = $form->title;
-		}
-
-		return $choices;
-	}
-
 
 }