+++ wp-content/plugins/core/src/Blocks/Types/Lead_Form/Lead_Form_Model.php
@@ -1,5 +1,4 @@
-<?php
-declare( strict_types=1 );
+<?php declare( strict_types=1 );
 
 namespace Tribe\Project\Blocks\Types\Lead_Form;
 
@@ -15,13 +14,15 @@ class Lead_Form_Model extends Base_Model {
 		return [
 			Lead_Form_Block_Controller::ATTRS       => $this->get_attrs(),
 			Lead_Form_Block_Controller::CLASSES     => $this->get_classes(),
-			Lead_Form_Block_Controller::LAYOUT      => $this->get( Lead_Form::LAYOUT, Lead_Form::LAYOUT_CENTER ),
+			Lead_Form_Block_Controller::LAYOUT      => $this->get( Lead_Form::LAYOUT, Lead_Form::LAYOUT_BOTTOM ),
 			Lead_Form_Block_Controller::WIDTH       => $this->get( Lead_Form::WIDTH, Lead_Form::WIDTH_GRID ),
 			Lead_Form_Block_Controller::TITLE       => $this->get( Lead_Form::TITLE, '' ),
 			Lead_Form_Block_Controller::LEADIN      => $this->get( Lead_Form::LEAD_IN, '' ),
 			Lead_Form_Block_Controller::DESCRIPTION => $this->get( Lead_Form::DESCRIPTION, '' ),
 			Lead_Form_Block_Controller::CTA         => $this->get_cta_args(),
-			Lead_Form_Block_Controller::FORM        => (int) $this->get( Lead_Form::FORM, 0 ),
+			Lead_Form_Block_Controller::BACKGROUND  => $this->get( Lead_Form::BACKGROUND, Lead_Form::BACKGROUND_LIGHT ),
+			Lead_Form_Block_Controller::FORM_FIELDS => $this->get( Lead_Form::FORM_FIELDS, Lead_Form::FORM_STACKED ),
+
 		];
 	}