+++ wp-content/themes/core/components/document/header/header.php
@@ -25,7 +25,7 @@ $c = Header_Controller::factory();
 	<?php // PLATFORM META: IE ?>
 	<meta name="application-name" content="<?php bloginfo( 'name' ); ?>">
 
-	<?php do_action( 'wp_head' ) ?>
+	<?php do_action( 'wp_head' ); ?>
 
 </head>
 
@@ -34,5 +34,9 @@ $c = Header_Controller::factory();
 	<?php do_action( 'tribe/body_opening_tag' ) ?>
 
 	<div class="l-wrapper" data-js="site-wrap">
+
+		<a href="#main-content" class="a11y-skip-link u-visually-hidden">
+			<?php esc_html_e( 'Skip to main content', 'tribe' ); ?>
+		</a>
 
 		<?php get_template_part( 'components/header/masthead/masthead' ); ?>